% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visibility_graph.R
\name{visibility_graph}
\alias{visibility_graph}
\title{visibility_graph}
\usage{
visibility_graph(time_series)
}
\arguments{
\item{time_series}{time series}
}
\value{
Visibility Graph as igraph object.)
}
\description{
Create Visibility Graph
}
