% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence_network.R
\name{recurrence_matrix}
\alias{recurrence_matrix}
\title{Recurrence matrix}
\usage{
recurrence_matrix(ts, eps)
}
\value{
Recurrence matrix
}
\description{
Create recurrence matrix
}
\examples{
recurrence_matrix(rnorm(10), 0.5)
}
