% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence_network.R
\name{recurrence_matrix}
\alias{recurrence_matrix}
\title{recurrence_matrix}
\usage{
recurrence_matrix(ts, eps)
}
\arguments{
\item{ts}{time_series.}

\item{eps}{lag}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
recurrence_matrix
}
