% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrence_network.R
\name{recurrence_network}
\alias{recurrence_network}
\title{recurrence_network}
\usage{
recurrence_network(ts, eps)
}
\value{
Recurrence network as igraph object.
}
\description{
Create recurrence network
}
