% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_network.R
\name{transition_network}
\alias{transition_network}
\title{transition_network}
\usage{
transition_network(ts, lag, embedding)
}
\arguments{
\item{lag}{lag}

\item{embedding}{embds}

\item{time_series}{time_series.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
transition_network
}
