% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_network.R
\name{transition_network}
\alias{transition_network}
\title{transition_network}
\usage{
transition_network(ts, lag, embedding)
}
\value{
Transition network as igraph object.
}
\description{
Create transition network
}
